Friends List Part 2 - Leaderboard & User Metadata
Rework of SampleMainMenu - Part 2.
-Moved Leaderboard (not scoreboard) from SampleMainMenu.cs to Separate Class.
-Added an event to inform listeners once the Connector.instance has finished loading everything (leaderboard was attempting to publish before connector had finished).
-Moved Metadata loading to Connector.cs to ensure all usermetadata is fully loaded before allowing users to create or join a room.
-Updated PNManager.GetUserMetadata to create and set user metadata in case the user has previously not registered (first time opening app).
-Load metadata setup for user, including hats, language, and fps option.
-Implement PNObject listener in Connector. Created delegate to signal other listeners. Listening for user metadata updates in sample main menu (Not choosing to do so in Connector.cs, but certainly plausable).
-
