+++ Assets/SuperMultiplayerShooter/Scripts/Leaderboard.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3c7d3c83b290fea44b5fd72c8607f1ae
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
